% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_tables.R
\name{pgn_acousticvalues}
\alias{pgn_acousticvalues}
\title{Accustic values}
\usage{
pgn_acousticvalues(con)
}
\arguments{
\item{con}{A connection to pgnapes}
}
\value{
An sql tibble
}
\description{
Accustic values
}
