% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_connect.R
\name{pgn_connect}
\alias{pgn_connect}
\title{Connect to PGNAPES}
\usage{
pgn_connect(username, password, use_odbc = TRUE)
}
\arguments{
\item{username}{your username}

\item{password}{your password}

\item{use_odbc}{A boolean (default TRUE) controlling if connection via odbc or ROracle.
The latter is generally faster but more difficult to setup.}
}
\value{
a database connection
}
\description{
This function is the mother of all mothers that all other ft_ functions rely on. If not you are doomed.
}
