% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_tables.R
\name{pgn_icessquares_big}
\alias{pgn_icessquares_big}
\title{icessquares_big}
\usage{
pgn_icessquares_big(con)
}
\arguments{
\item{con}{A connection to pgnapes}
}
\value{
An sql tibble
}
\description{
icessquares_big
}
