% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_tables.R
\name{pgn_tbl}
\alias{pgn_tbl}
\title{A convenient function}
\usage{
pgn_tbl(con, table)
}
\arguments{
\item{con}{A connection to pgnapes}

\item{table}{A character specifying the name of the table}
}
\value{
An sql tibble
}
\description{
A convenient function
}
