% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_connect.R
\name{pgn_connect}
\alias{pgn_connect}
\title{Connect to PGNAPES}
\usage{
pgn_connect(username, password)
}
\arguments{
\item{username}{your username}

\item{password}{your password}
}
\value{
a database connection
}
\description{
Establish a connection to PGNAPES
}
